LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCPeerConnection-connectionState.https-expected.txt


Harness Error (TIMEOUT), message = null

PASS Initial connectionState should be new
PASS Closing the connection should set connectionState to closed
TIMEOUT connection with one data channel should eventually have connected connection state Test timed out
TIMEOUT connection with one data channel should eventually have transports in connected state Test timed out
FAIL connectionState remains new when not adding remote ice candidates assert_equals: expected "new" but got "checking"
TIMEOUT connectionState transitions to connected via connecting Test timed out
NOTRUN Closing a PeerConnection should not fire connectionstatechange event
